LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_disabled_form_control-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: pointerevents fire to disabled form controls

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: Pointerevents are dispatched to disabled form controls.

Use the mouse or pen to press on the purple button. Or with touch, tap on the purple box.
Disabled button

Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse pointerevent attributes
